+++ SRC/Model/Service/ArService.php
@@ -1,0 +1,15 @@
+<?php 
+$path = dirname(__DIR__, 1);
+    include_once($path."/DAO/ArDAO.php");
+    function Feltolto($kapcsolat){
+        return FajtaFeltolto($kapcsolat);
+    }
+    function Kategoria($kapcsolat){
+        return KategoriaFeltolto($kapcsolat);
+    }
+    function KornyezetVedelem($kapcsolat){
+        return KornyezetVedelemFeltolto($kapcsolat);
+    }
+    function AutotipusAdatAtvevo($marka,$tipus,$fajta,$kategoria,$premium,$kornyezetvedelem){
+       return AutoTipusTarolo($marka,$tipus,$fajta,$kategoria,$premium,$kornyezetvedelem);
+    }
