+++ SRC/View/Arfelvetel.php
@@ -1,0 +1,53 @@
+<!DOCTYPE html>
+<html lang="hu">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Árfelvevő</title>
+    <?php
+    include_once("../Controller/ArFelvevoController.php");
+    ?>
+</head>
+<body>
+    <form action="" method="POST">
+        <label name="marka" for="">
+            <input type="text" name="marka" required placeholder="Márka">
+        </label>
+        <label name="tipus" for="">
+            <input type="text" name="tipus" required placeholder="Tipus">
+        </label>
+        <select name="fajta" id="" required>
+            <option value="">Válaszon fajtát</option>
+            <?php
+            getFajta($kapcsolat);
+            ?>
+        </select>
+        <select name="kategoria" required id="">
+            <option value="">Válaszon kategoriát</option>
+            <?php
+                getKategoria($kapcsolat);
+            ?>
+        </select>
+        <label for="">
+            <input name="premium" type="checkbox" value="off">
+                Prémium
+            </input>
+        </label>
+        <select name="kornyezetvedelem" required id="">
+            <option value="">Válaszon környezetvédelmi besorolást</option>
+            <?php
+                getKornyezetVedelem($kapcsolat);
+            ?>
+        </select>
+            <?php
+                initAutotipusbekuldes();
+            ?>
+        <button type="submit" name="Autotipusbekuldes">Beküldés</button>
+    </form>
+    <div>
+        <?php
+        printresult();
+        ?>
+    </div>
+</body>
+</html>
